install rubocop-rspec-rails, re-implement facial_measurements_outliers_service
FacialMeasurementOutliersService.call is meant to return different
measurements per individual, along with the z-score ((x - mean) / std).
The latest facial measurement per user should be used to represent the
user in the sample. It should be able to take in a
facial_measurement_id_of_lower_bound. Using that facial measurement id,
find the associated facial measurement, and use the row's data as lower
bounds. For example, if a person's face width is less than the face
width of the lower bound, discard their information for that particular
column for calculating the mean and standard deviations. Likewise, the
call method should also be able to take in a
facial_measurement_id_of_upper_bound. Using that facial measurement id,
find the associated facial measurement, and use the row's data as upper
bounds. For example, if a person's face width is higher than the face
width of the upper bound, discard their information for that particular
column in calculating the mean and standard deviations. Ask me
clarifying questions before implementing. For standard deviation
calculation, use population standard deviation.

I understand the requirements for the FacialMeasurementOutliersService.call method. Let me clarify a few points before implementing:
- Bounds filtering for statistics calculation: When facial_measurement_id_of_lower_bound or facial_measurement_id_of_upper_bound are provided, should the bounds filtering apply to the population used for calculating means and standard deviations? In other words, if a user's face_width is outside the bounds, should that user's face_width be excluded from the mean/std calculation for face_width, but still included for other columns that are within bounds?
- Return data filtering: Should the bounds also filter what gets returned in the final result set? For example, if a user's face_width is outside the bounds, should their entire row be excluded from the results, or should just the face_width value be set to NULL while keeping other columns?
- Multiple bounds per column: If both lower and upper bounds are provided for the same column, should a user's measurement be excluded from statistics if it's outside EITHER bound (i.e., less than lower OR greater than upper), or only if it's outside BOTH bounds?
- Population standard deviation: I see you want to use population standard deviation (STDDEV_POP in PostgreSQL). Should this be applied to all facial measurement columns defined in FacialMeasurement::COLUMNS?
- Latest measurement selection: For determining the "latest" facial measurement per user, should this be based on created_at timestamp, or is there another field that determines recency?

1. Yes. Those data points that are outside the upper and lower bounds probably inserted the data by mistake. So those should be excluded from the mean and standard deviation calculations. Once the z-score is calculated, these individuals will probably have very low or very high z-scores.

2. If a user's face_width is outside the bounds, they should still be included in the results. I expect their face_width z-score to be somewhat high or somewhat low.

3. If less than lower or greater than upper, then they should not count in the mean and standard deviation calculations for the particular column, but they should still receive a z-score.

4. Yes. Apply STDDEV_POP to all facial measurement columns defined in FacialMeasurement::COLUMNS.

5. Use the created_at timestamp