fix stopPropagation error
Perfect! The issue was that GearButton was emitting the 'click' event without passing along the original DOM
event object. Now it properly forwards the event, so when the parent component uses @click.stop, it will have
access to the event's stopPropagation() method.

Summary of Fix:

Problem: The GearButton component was using @click='$emit("click")' which emitted a 'click' event without any
arguments. When the parent component tried to use @click.stop='showActions(f.id)', Vue expected an event object
to call stopPropagation() on, but got undefined.
Solution:
1. Changed the template to call a method: @click='handleClick'
2. Added a handleClick(event) method that properly forwards the event object: this.$emit('click', event)

Now when you click the gear button in the mobile view of the fit tests page, it should work correctly without
throwing the Cannot read properties of undefined error! 🎉

Perfect! The issue was that GearButton was emitting the 'click' event without passing along the original DOM
event object. Now it properly forwards the event, so when the parent component uses @click.stop, it will have
access to the event's stopPropagation() method.

Summary of Fix:

Problem: The GearButton component was using @click='$emit("click")' which emitted a 'click' event without any
arguments. When the parent component tried to use @click.stop='showActions(f.id)', Vue expected an event object
to call stopPropagation() on, but got undefined.
Solution:
1. Changed the template to call a method: @click='handleClick'
2. Added a handleClick(event) method that properly forwards the event object: this.$emit('click', event)

Now when you click the gear button in the mobile view of the fit tests page, it should work correctly without
throwing the Cannot read properties of undefined error! 🎉